Implement attribute partitioning methods with better guarantees
Le coclustering se base sur la generation de solutions voisines d'une solution donnee
en sur-partitionnant les attributs avec des parties additionnelles.
Les methodes actuellesment implementee dans KWDataGridManager sont limitees
par les heuristiques utilisees:
- nombres de partie obtenues incorrectes
- equilibage (en effectif) des partie controle uniquement bia la granularisation ou la tokensation coutante

Il s'agit d'implementer de nouvelle methodes, avec des garanties accrues:
- on obtient exactement le nombre de parties demandees, si exige
- nouveau parametre permettant de controler le taux de d'equilibre des partitions
  - permet de generer ue partition au plus proche d'une partition en frequence egale
  - ou de tolerer de facon controlee des partie d'effectif moindre

Cela permettra de rendre les generations de solution voisine plus efficace.

KWDataGridManager::AddContinuousAttributeRandomParts
- Parametres
  - sourceAttribute,
  - mandatoryAttribute
  - targetAttribute
  - nRequestedPartNumber
  - nMinimimEqualFrequencyPartNumber
  - bForce
- Documenattion
  // Export aleatoire de parties pour un attribut continu donne.
  // La methode construit d'abord une discretisation de reference en M
  // intervalles quasi equi-effectifs (M = nMinimimEqualFrequencyPartNumber),
  // puis choisit aleatoirement des bornes dans cette grille pour produire
  // R parties (R = nRequestedPartNumber).
  //
  // Si mandatoryAttribute est present, les nouvelles parties sont obtenues
  // par sur-partition de ses parties existantes: les bornes candidates
  // doivent respecter les bornes de mandatoryAttribute.
  //
  // Si R = M, la partition est la plus equilibree possible.
  // Si R < M, la partition peut etre moins equilibree (plus de variabilite).
  //
  // Si le nombre de valeurs distinctes est insuffisant, le nombre de parties
  // obtenu peut etre inferieur a R. Avec bForce=false, on accepte ce resultat.
  // Avec bForce=true, on tente d'atteindre R en relachant progressivement
  // la contrainte d equilibre, tout en respectant mandatoryAttribute s il existe.

KWDataGridManager::ExportGroupableAttributeRandomParts
  // Export aleatoire de parties pour un attribut categoriel donne.
  // Comportement similaire a celui du cas d'un attribut numerique

Algorithmes:
- utilisation de QuantileBuilder dans les deux cas
- nouvelle classe technique interne KWDGMGroupableAttributePartInformation dans le cas des attribut groupables
- attention particuliere a la complexite algorithmique, en O(n log n)
  - y compris dans le cas Continuous, avec exploration systelatique des nombre de quantiles rendus par le quantileBuilder

Autres modifications mineures
- KWDataGrid: passage en const des method exploitant les structures d'indexation (qui elles sont mutables)
- KWDataGridManager: passage en const des parametres mandatoryAttribute des methdoes concernees

+++ src/Learning/KWDataPreparation/KWDataGridOptimizer.cpp
@@ -448,7 +448,7 @@ void KWDataGridOptimizer::GenerateNeighbourSolution(const KWDataGrid* initialDat
 
 	// Calcul du nombre de parties a exporter, en fonction du niveau de bruit et du nombre moyen de partie existantes
 	nMeanOptimizedPartNumber =
-	    ceil(optimizedDataGrid->GetTotalPartNumber() / (double)optimizedDataGrid->GetAttributeNumber());
+	    (int)ceil(optimizedDataGrid->GetTotalPartNumber() / (double)optimizedDataGrid->GetAttributeNumber());
 	nRequestedContinuousPartNumber =
 	    1 + nMeanOptimizedPartNumber + (int)ceil(dNoiseRate * nMaxContinuousPartNumber);
 	nRequestedSymbolPartNumber = 1 + nMeanOptimizedPartNumber + (int)ceil(dNoiseRate * nMaxSymbolPartNumber);
